requirements.txt
Flask
requests
beautifulsoup4
lxml
python-dotenv
gunicorn
tzdata
supabase
stripe
python-jose[cryptography]
flask-limiter
